# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Survey Research Automation",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/research/jobs": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Create a research job",
        "description": "Starts auto-fetching documents for an order. Accepts a list of document types (from the source registry). Returns the created job with per-document status.",
        "operationId": "create_research_job_api_v1_research_jobs_post",
        "parameters": [
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResearchJobRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_ResearchJob_"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/research": {
      "post": {
        "summary": "Research",
        "operationId": "research_api_research_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs": {
      "get": {
        "summary": "Get Jobs",
        "operationId": "get_jobs_api_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Job",
        "description": "Delete a saved job/search record (its whole folder).",
        "operationId": "delete_job_api_jobs_delete",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/feedback": {
      "post": {
        "summary": "Post Feedback",
        "description": "Record the surveyor's match/mismatch verdict (+ optional note) on a fetched document.",
        "operationId": "post_feedback_api_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/upload": {
      "post": {
        "summary": "Upload Doc",
        "description": "Attach a manually-downloaded document to a job's doc step (one upload per doc; replaces\nany previous one and syncs the locker). Returns {file, removed:[...]}.",
        "operationId": "upload_doc_api_jobs_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_doc_api_jobs_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/doc": {
      "delete": {
        "summary": "Delete Job Doc",
        "description": "Delete one document file from a job (and remove it from its Evidence Locker).",
        "operationId": "delete_job_doc_api_jobs_doc_delete",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Name"
            }
          },
          {
            "name": "file",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "File"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evidence/send": {
      "post": {
        "summary": "Evidence Send",
        "description": "Send selected job documents to the order's Evidence Locker.",
        "operationId": "evidence_send_api_evidence_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvidenceSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evidence": {
      "get": {
        "summary": "Evidence List",
        "operationId": "evidence_list_api_evidence_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Evidence Delete Order",
        "description": "Delete an entire Evidence Locker order (folder + all its documents).",
        "operationId": "evidence_delete_order_api_evidence_delete",
        "parameters": [
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evidence/detail": {
      "get": {
        "summary": "Evidence Detail",
        "operationId": "evidence_detail_api_evidence_detail_get",
        "parameters": [
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evidence/item": {
      "delete": {
        "summary": "Evidence Delete Item",
        "description": "Remove one document from an order's Evidence Locker.",
        "operationId": "evidence_delete_item_api_evidence_item_delete",
        "parameters": [
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Order"
            }
          },
          {
            "name": "file",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "File"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateResearchJobRequest": {
        "properties": {
          "order_id": {
            "type": "string",
            "format": "uuid",
            "title": "Order Id",
            "description": "Order to research"
          },
          "document_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Document Types",
            "description": "Document types to fetch, e.g. ['PARCEL_RECORD', 'FLOOD']. Empty = fetch all applicable types for the order's survey type."
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key",
            "description": "Client-generated idempotency key. Auto-generated if not provided."
          }
        },
        "type": "object",
        "required": [
          "order_id"
        ],
        "title": "CreateResearchJobRequest",
        "description": "Request to start a research job for an order.\n\n``document_types`` is the list of document types to fetch (from the\nsource registry). If empty, all applicable types are fetched.\n\n``idempotency_key`` is optional; if omitted, the server generates one.\nDuplicate requests with the same key return the existing job."
      },
      "DataEnvelope_ResearchJob_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ResearchJob"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataEnvelope[ResearchJob]"
      },
...
    }
  }
}